Create The for Loop
the formula of for loop is: for (initialization; condition; increment) statement;
However the code count = count+1 you will not see in the professional C++ writing.
This code basically makes the computer start to count from 1 to 100. You can change the count to any number. 
P.S all this is for the new programmers or people who try to write a code.