Create Using Blocks of Code
This code will let you divide numbers (only divide). If you need a fully proper working calculator, you can use the one that i made.  The link to it : https://github.com/Max634/Basic-Calculator-V-2.0#readme
The whole point of this is to learn how to use if statements in the code.